One step on the way to reverting back to how caimanc handles command lists
encoding order needs to match submission order, for now, because of how
wgpu handles Queue::write_buffer

There's no inter-commandlist synchonization and no parallel submission
or multiple queues, so there really isn't much of a point to having
multiple encoders active like there is in Vulkan.  I keep making the
mistake of thinking this capability is useful.  It might eventually be,
but not right now.