Start Init Algorithm Part 1
The initialization algorithm has 2 parts. The first part
places initializations at the latest possible points. This
change begins work on this algorithm by introducing the
init_synth module which contains functions to generate
initializer sets of each uninitialized mutable variable.

An initializer set is a set of hole locations that collectively
dominate the uses of the mutable they initialize.