Minimize initializer sets
To prepare for the second phase of the frontend initialization
algorithm, which hoists the initialization points to better locations,
if any exist, we first minimize the initialization set by removing
hole locations that are collectively strict dominated by a subset
of the initialization set. That is to say, the mutable must already
be initialized at the program point, and so it can be removed without
affecting correctness.