Don't assume phi nodes have select quotients
This change enables writes to variables that don't have
select quotients to occur withen the branches of an if.
This is done by introducing a more flexible SpliceTerm
type expression that can either match like a regular
term or replace itself with a child, provided all its
children can unify with eachother.

Further, we allow phi nodes to unify with constraints
that aren't selects by making the phi node constraints
to be a SpliceTerm.

The coding pattern this change enables is a bit strange
to write manually, but can be easily generated by the
frontend by putting holes in both branches of an if.

Finally, this change also introduces some utilities needed
for phase 2 of the initialization algorithm.