Initializer Placement Phase 2
This implements a mostly working hoist algorithm to improve upon
the placement of initializations of variables. At a high-level,
we hoist an initialization point to a set of collectively
dominating holes if we must initialize less dependencies at those
holes. To take into account sharing of dependencies between other
variables, a given variable is only considered to initialize `1/n`
of a node where `n` is the number of other initializations at the
hole that also depend on the same node.